Create Peter
I have the Makerbase MKS DLC32 V2.1 CNC motherboard, I cannot make a connection to USB. Have the older MKS board and works OK. Need to use it for CNC Routing. What do i need to do to get it to connect. Have the CH340 installed and working with the other board. regards.